Split auto-languageclient into two parts
The point of this excercise is to isolate adapters to a single part of
the code base, and make the rest completely adapter-independent.
