+++ lib/main.ts
@@ -4,6 +4,7 @@
 // tslint:enable:no-reference
 
 import AutoLanguageClient from './auto-languageclient';
+import BaseLanguageClient from './base-languageclient';
 import Convert from './convert';
 import DownloadFile from './download-file';
 import LinterPushV2Adapter from './adapters/linter-push-v2-adapter';
@@ -11,6 +12,7 @@ import LinterPushV2Adapter from './adapters/linter-push-v2-adapter';
 export * from './auto-languageclient';
 export {
   AutoLanguageClient,
+  BaseLanguageClient,
   Convert,
   DownloadFile,
   LinterPushV2Adapter,
